Support 2-character chain identifiers
I am using Bio.PDB to parse structures in mmCIF and PDB format. I realised that PDBIO does not deal well with two-character chain identifiers (like ‘AA’ or ‘AB’) found in assembly structures. What this pull request basically does is checking the length of the chain identifier string and adding a space in front of it, if is a single character. The formatting string is modified accordingly.